refactor: refactor the folder structure and create a public API file
