+++ packages/analytics_manager/lib/src/firebase_analytics_manager.dart
@@ -2,7 +2,7 @@ import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:log_manager/log_manager.dart';
 
-import '../analytics_manager.dart';
+import 'analytics_manager.dart';
 
 /// A manager class that handles analytics using Firebase Analytics.
 /// This class implements the [AnalyticsManager] interface and
